chore: rework test structure
