--- packages/integration-karma/test/light-dom/assigned-node-element/x/shadowContainer/shadowContainer.html
@@ -1,3 +1,0 @@
-<template>
-    <x-light-container></x-light-container>
-</template>
