+++ packages/integration-karma/test/light-dom/synthetic-shadow-styles/x/container/container.js
