+++ packages/integration-karma/test/light-dom/synthetic-shadow-styles/x/grandchild/grandchild.css
