+++ packages/integration-karma/test/light-dom/synthetic-shadow-styles/x/light/light.js
