+++ packages/integration-karma/test/light-dom/synthetic-shadow/x/consumer/consumer.js
