+++ packages/integration-karma/test/light-dom/synthetic-shadow/x/lightContainer/lightContainer.js
