+++ packages/integration-karma/test/light-dom/synthetic-shadow/x/shadowContainer/shadowContainer.html
@@ -1,0 +1,3 @@
+<template>
+    <x-light-container data-id="light-container"></x-light-container>
+</template>
