+++ packages/integration-karma/test/light-dom/synthetic-shadow/x/shadowContainer/shadowContainer.js
